4th part done.

+++ src/main/java/se/lexicon/ahmad/abdou/booklender/entity/Book.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int bookId;
+    private long bookId;
     private String title;
     private boolean available;
     private boolean reserved;
